Filetask dhtmlx

+++ src/main/java/com/example/filetask/service/UserService.java
@@ -37,15 +37,20 @@ public class UserService {
 
 
 
+    //List : 순서가 있으며, 데이터(값) 중복 허용
+    //Set : 순서가 없으며, 데이터(값) 중복을 허용하지 않음
+    //Map : Key&Value 구조, Key는 중복을 허용하지 않으며, Value(값)은 중복을 허용
+    //파일 업로드 처리를 하고, 그 결과(성공 여부, 파일명 등)를 Map 형태의 데이터 묶음으로 반환하는 함수.
     public Map<String, Object> uploadFile(MultipartFile file, boolean force) throws IOException {
 
         Map<String, Object> result = new HashMap<>();
         String fileName = file.getOriginalFilename();
 
+        //만약에 JS에서 뚫리면 이걸로 잡음
         if (fileName == null || !fileName.endsWith(".dbfile")) {
             throw new RuntimeException("dbfile 파일만 업로드할 수 있습니다.");
-
         }
+
         String redisKey = "recent:" + fileName;
         Boolean duplicated = redisTemplate.hasKey(redisKey);
 
@@ -64,6 +69,7 @@ public class UserService {
             return result;
         }
 
+        //여기서부터 시작 jsp -> PageController -> JS(업로드 버튼) -> 후에 시작하는곳
         BufferedReader br = new BufferedReader(new InputStreamReader(file.getInputStream()));
 
         List<String> lines = new ArrayList<>();
@@ -131,9 +137,6 @@ public class UserService {
         return userQueryRepository.searchUsers(field, keyword);
     }
 
-
-
-
     public void signup(Info user) {
         if(infoRepository.existsById(user.getId())){
             throw new RuntimeException("존재하는 아이디입니다.");
@@ -142,9 +145,8 @@ public class UserService {
         infoRepository.save(signupUser);
     }
 
-
     public Info login(String id, String pwd) {
-        Info user = infoRepository.findById(id).orElseThrow(()-> new RuntimeException("아이디가 없다."));
+        Info user = infoRepository.findById(id).orElseThrow(()-> new RuntimeException("아이디가 없습니다."));
 
         if(!user.getPwd().equals(pwd)){
             throw new RuntimeException("비밀번호가 없습니다");
@@ -153,8 +155,14 @@ public class UserService {
         return user;
     }
 
+
     //Swagger용
+    //ResponseEntity
     public ResponseEntity<String> deleteById(String id) {
+        if (!userRepository.existsById(id)) {
+            return ResponseEntity.status(org.springframework.http.HttpStatus.NOT_FOUND).body("NOT_FOUND");
+        }
+
         userRepository.deleteById(id);
         return ResponseEntity.ok("삭제");
     }
